+++ src/oxide/paint/hitbox.rs
@@ -61,7 +61,7 @@ impl Paint {
             }
             let team = vmt_call!(pipe, get_team_number);
 
-            let hitbox = pipe.get_hitbox(HitboxId::Head)?;
+            let hitbox = pipe.get_hitbox(HitboxId::Head as usize)?;
             self.draw_hitbox(frame, hitbox, team.color(), 10)?;
         }
         Ok(())
